fix: BigDecimal for NUMERIC columns, ddl-auto none
- Change averageRating Double → BigDecimal in entity, DTO, repo, service, controller
  (Hibernate 6 maps Double to float8, but DB column is NUMERIC — type mismatch)
- Change minRating Double → BigDecimal in SearchRequest and SearchController
- Set ddl-auto: none (Flyway owns schema management, Hibernate must not validate/alter it)
- Update mobile type: averageRating as string (BigDecimal serialises to string in JSON)

+++ service-contractor/src/main/java/com/kerjapro/contractor/controller/SearchController.java
@@ -11,6 +11,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/api/contractors/search")
 @RequiredArgsConstructor
@@ -26,7 +28,7 @@ public class SearchController {
             @RequestParam(required = false) String brandName,
             @RequestParam(required = false) String city,
             @RequestParam(required = false) String state,
-            @RequestParam(required = false) Double minRating,
+            @RequestParam(required = false) BigDecimal minRating,
             @RequestParam(required = false) Boolean verified,
             @RequestParam(defaultValue = "0")  int page,
             @RequestParam(defaultValue = "20") int size,
